requirements.txt
Flask
flask_sqlalchemy
flask_marshmallow
marshmallow-sqlalchemy
